refactor: Move function decl building to its own component
