+++ compiler/src/CMakeLists.txt
@@ -26,6 +26,7 @@ add_library(
 set(FLUIR_COMPILER_FRONTEND_SOURCES
     "frontend/parser.cpp"
     "frontend/ast_builder.cpp"
+    "frontend/ast_builder/function_builder.cpp"
     "frontend/text_parse.cpp"
     "frontend/type_checker.cpp"
 )
